Flexible/responsive multi-column layout for upgrades.
The huge list of upgrades now uses CSS3 `columns` property, dynamically
adjusting the number of columns according to the available space.

Also, the upgrades are now ordered from the top to the bottom, then to
the next column. This way it is much easier to read.

Also added a custom style to checkboxes for upgrades and for managers.
For now, it just strikes through the text if such upgrade has been
bought. Suggestions for better styles are welcome.

Minor stuff:

* Changed one width from pixels to `em`.
* Added background-color mouse-over to `<label>`.
* Made the selector for the tabs more specific (in case someday in the
  future we want to use `<li>` for other purpose).