Fixed Angels Upgrade Suggestions
Problem: if the user typed an angel count far above the lowest angel
upgrade (say from 1 tretrigintillion to 1 duoquadragintlillion) and
then changed it back to a lower value, the angel upgrades above the new
angel count were not dismissed as “exclamation” upgrades. This fix
removes the alert from any that are above the angel count.