Update Profit-A-Bowl
Adds "gilding"; corrects baseCost and basePower for "Fake Fans". The prices for "Fake Fans" were off by a few cents early on which led me to believe something was a just a titch off.

It seemed likely the basePower would be a "nice" number and baseCost would be either a nice number in and of itself or be a nice number related to basePower.

Since the initial purchase price of one "Fake Fans" was $5, I tried setting basePower to 1.007 (instead of 1.006999855) and baseCost to 5/1.0007 (4.996502448). That made the costs match up to reality more closely.

With the old settings, buying 100 more "Fake Fans" when owning 200 was priced at 595.08 but the calculator projected it to be 595.50; with the new settings, it's projected to be 595.08(321).